% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{matrixToVec}
\alias{matrixToVec}
\title{matrixToVec}
\usage{
matrixToVec(X)
}
\arguments{
\item{X}{a SYMETRIC matrix}
}
\value{
a vector  contenting the coefficient of the upper triangle of the matrix X
from left to right and from top to bottom.
}
\description{
matrixToVec
}
