% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{plotGraphs}
\alias{plotGraphs}
\title{plot the data matrix, the inferred graph and/or the true binary graph}
\usage{
plotGraphs(dataMatrix = NULL, inferredGraph = NULL, binaryTruth = NULL)
}
\arguments{
\item{dataMatrix}{observed data matrix}

\item{inferredGraph}{graph inferred by the multiple testing procedure via graphInference()}

\item{binaryTruth}{true binary graph}
}
\value{
a list of FDR and TDR values, if possible
}
\description{
plot the data matrix, the inferred graph and/or the true binary graph
}
