% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{rnsbm}
\alias{rnsbm}
\title{return a random NSBM}
\usage{
rnsbm(p, theta, modelFamily = "Gauss")
}
\arguments{
\item{p}{(interger) number of node in the network}

\item{theta}{=(pi;w;nu0;nu) parameter of the model}

\item{modelFamily}{the distribution family of the noise under the null hypothesis,
which can be "Gauss" (Gaussian), "Gamma", or "Poisson", by default it's 'Gauss'}
}
\value{
X the noisy matrix

theta

latentZ the latent clustering

latentA the latent adjacency matrix
latent variables
we strat by sampling the latent variable Z which is the vector containing the family of each nodes
adjacency matrix
then we sample the adjacency matrix, conditionally to Z the coordinate of A follow a binomial
of a parameter contain in theta$w
noisy observations under the null
we create a matrix (n,n) X and we initialize all its entry (half of them is undirected)
with a sampling of the law under the null
then for each entry where A is none zero we sample it according to the law under the alternative
}
\description{
return a random NSBM
}
