% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{vecToMatrix}
\alias{vecToMatrix}
\title{vecToMatrix}
\usage{
vecToMatrix(X, p)
}
\arguments{
\item{X}{a vector}

\item{p}{(integer) the dimension of the square matrix returned by the function
be careful the length of the vector X must be equal to p(p+1)/2}
}
\value{
a p by p symetric matrix whose upper triangle coefficients from left to right and
from top to bottom are the entries of the vector X
}
\description{
vecToMatrix
}
